agentic-tagging-workbench-backend-main/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
pandas==2.1.4
openpyxl==3.1.2
pillow==10.2.0
python-dotenv==1.0.0
tiktoken==0.5.2
openai>=1.56.0
playwright==1.46.0
requests==2.32.3
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.5.0
python-multipart==0.0.6
